fixed int to long fromid

+++ Binance.Net/BinanceClient.cs
@@ -244,7 +244,7 @@ namespace Binance.Net
         /// <param name="endTime">Time to stop getting trades from</param>
         /// <param name="limit">Max number of results</param>
         /// <returns>The aggregated trades list for the symbol</returns>
-        public CallResult<BinanceAggregatedTrades[]> GetAggregatedTrades(string symbol, int? fromId = null, DateTime? startTime = null, DateTime? endTime = null, int? limit = null) => GetAggregatedTradesAsync(symbol, fromId, startTime, endTime, limit).Result;
+        public CallResult<BinanceAggregatedTrades[]> GetAggregatedTrades(string symbol, long? fromId = null, DateTime? startTime = null, DateTime? endTime = null, int? limit = null) => GetAggregatedTradesAsync(symbol, fromId, startTime, endTime, limit).Result;
 
         /// <summary>
         /// Gets compressed, aggregate trades. Trades that fill at the time, from the same order, with the same price will have the quantity aggregated.
@@ -255,7 +255,7 @@ namespace Binance.Net
         /// <param name="endTime">Time to stop getting trades from</param>
         /// <param name="limit">Max number of results</param>
         /// <returns>The aggregated trades list for the symbol</returns>
-        public async Task<CallResult<BinanceAggregatedTrades[]>> GetAggregatedTradesAsync(string symbol, int? fromId = null, DateTime? startTime = null, DateTime? endTime = null, int? limit = null)
+        public async Task<CallResult<BinanceAggregatedTrades[]>> GetAggregatedTradesAsync(string symbol, long? fromId = null, DateTime? startTime = null, DateTime? endTime = null, int? limit = null)
         {
             var parameters = new Dictionary<string, object> { { "symbol", symbol } };
             parameters.AddOptionalParameter("fromId", fromId?.ToString());
